+++ src/Mathematics.NET/DifferentialGeometry/AutoDiffTensor2`3.cs
@@ -33,24 +33,22 @@ using Mathematics.NET.DifferentialGeometry.Abstractions;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a rank-one tensor of two variables for use in forward-mode automatic differentiation.</summary>
-/// <typeparam name="TDN">A type that implements <see cref="IDual{T, U, V, W}"/>.</typeparam>
-/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
+/// <summary>Represents a rank-one tensor of two variables for use in reverse-mode automatic differentiation.</summary>
+/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
 /// <typeparam name="TI">An index.</typeparam>
-public record struct AutoDiffTensor2<TDN, TN, U, TI>
-    where TDN : IDual<TDN, TN, U, U>
-    where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
+public record struct AutoDiffTensor2<TN, U, TI>
+    where TN : IComplex<TN, U, U>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
     where TI : IIndex
 {
     /// <summary>The zeroth element of the rank-one tensor.</summary>
-    public TDN X0;
+    public Variable<TN, U> X0;
 
     /// <summary>The first element of the rank-one tensor.</summary>
-    public TDN X1;
+    public Variable<TN, U> X1;
 
-    public AutoDiffTensor2(TDN x0, TDN x1)
+    public AutoDiffTensor2(Variable<TN, U> x0, Variable<TN, U> x1)
     {
         X0 = x0;
         X1 = x1;
@@ -60,7 +58,10 @@ public record struct AutoDiffTensor2<TDN, TN, U, TI>
     // Indexer
     //
 
-    public TDN this[int index]
+    /// <summary>Get the element at the specified index.</summary>
+    /// <param name="index">An index.</param>
+    /// <returns>The element at the index.</returns>
+    public Variable<TN, U> this[int index]
     {
         readonly get => GetElement(this, index);
         set => this = WithElement(this, index, value);
@@ -68,7 +69,7 @@ public record struct AutoDiffTensor2<TDN, TN, U, TI>
 
     // Get
 
-    internal static TDN GetElement(AutoDiffTensor2<TDN, TN, U, TI> tensor, int index)
+    internal static Variable<TN, U> GetElement(AutoDiffTensor2<TN, U, TI> tensor, int index)
     {
         if ((uint)index >= 2)
             throw new IndexOutOfRangeException();
@@ -76,35 +77,29 @@ public record struct AutoDiffTensor2<TDN, TN, U, TI>
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static TDN GetElementUnsafe(ref AutoDiffTensor2<TDN, TN, U, TI> tensor, int index)
+    private static Variable<TN, U> GetElementUnsafe(ref AutoDiffTensor2<TN, U, TI> tensor, int index)
     {
         Debug.Assert(index is >= 0 and < 2);
-        return Unsafe.Add(ref Unsafe.As<AutoDiffTensor2<TDN, TN, U, TI>, TDN>(ref tensor), index);
+        return Unsafe.Add(ref Unsafe.As<AutoDiffTensor2<TN, U, TI>, Variable<TN, U>>(ref tensor), index);
     }
 
     // Set
 
-    internal static AutoDiffTensor2<TDN, TN, U, TI> WithElement(AutoDiffTensor2<TDN, TN, U, TI> tensor, int index, TDN value)
+    internal static AutoDiffTensor2<TN, U, TI> WithElement(AutoDiffTensor2<TN, U, TI> tensor, int index, Variable<TN, U> value)
     {
         if ((uint)index >= 2)
             throw new IndexOutOfRangeException();
-        AutoDiffTensor2<TDN, TN, U, TI> result = tensor;
+        AutoDiffTensor2<TN, U, TI> result = tensor;
         SetElementUnsafe(ref result, index, value);
         return result;
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static void SetElementUnsafe(ref AutoDiffTensor2<TDN, TN, U, TI> tensor, int index, TDN value)
+    private static void SetElementUnsafe(ref AutoDiffTensor2<TN, U, TI> tensor, int index, Variable<TN, U> value)
     {
         Debug.Assert(index is >= 0 and < 2);
-        Unsafe.Add(ref Unsafe.As<AutoDiffTensor2<TDN, TN, U, TI>, TDN>(ref tensor), index) = value;
+        Unsafe.Add(ref Unsafe.As<AutoDiffTensor2<TN, U, TI>, Variable<TN, U>>(ref tensor), index) = value;
     }
-
-    //
-    // Methods
-    //
-
-    public static AutoDiffTensor2<TDN, TN, U, TI> Create(TN x0, TN x1) => new(x0, x1);
 
     //
     // Formatting
@@ -116,7 +111,7 @@ public record struct AutoDiffTensor2<TDN, TN, U, TI>
     // Methods
     //
 
-    public static AutoDiffTensor2<TDN, TN, U, TNI> Create<TNI>(TDN x0, TDN x1)
+    public static AutoDiffTensor2<TN, U, TNI> Create<TNI>(Variable<TN, U> x0, Variable<TN, U> x1)
         where TNI : IIndex
         => new(x0, x1);
 }
