+++ src/Mathematics.NET/DifferentialGeometry/AutoDiffTensor3`4.cs
@@ -1,4 +1,4 @@
-// <copyright file="AutoDiffTensor2`2.cs" company="Mathematics.NET">
+// <copyright file="AutoDiffTensor3`4.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -33,35 +33,38 @@ using Mathematics.NET.DifferentialGeometry.Abstractions;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a rank-one tensor of two variables for use in reverse-mode automatic differentiation.</summary>
-/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/>.</typeparam>
+/// <summary>Represents a rank-one tensor of three variables for use in forward-mode automatic differentiation.</summary>
+/// <typeparam name="TDN">A type that implements <see cref="IDual{T, U, V, W}"/>.</typeparam>
+/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
 /// <typeparam name="TI">An index.</typeparam>
-public record struct AutoDiffTensor2<TN, U, TI>
-    where TN : IComplex<TN, U, U>
+public record struct AutoDiffTensor3<TDN, TN, U, TI>
+    where TDN : IDual<TDN, TN, U, U>
+    where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
     where TI : IIndex
 {
     /// <summary>The zeroth element of the rank-one tensor.</summary>
-    public Variable<TN, U> X0;
+    public TDN X0;
 
     /// <summary>The first element of the rank-one tensor.</summary>
-    public Variable<TN, U> X1;
+    public TDN X1;
 
-    public AutoDiffTensor2(Variable<TN, U> x0, Variable<TN, U> x1)
+    /// <summary>The second element of the rank-one tensor.</summary>
+    public TDN X2;
+
+    public AutoDiffTensor3(TDN x0, TDN x1, TDN x2)
     {
         X0 = x0;
         X1 = x1;
+        X2 = x2;
     }
 
     //
     // Indexer
     //
 
-    /// <summary>Get the element at the specified index.</summary>
-    /// <param name="index">An index.</param>
-    /// <returns>The element at the index.</returns>
-    public Variable<TN, U> this[int index]
+    public TDN this[int index]
     {
         readonly get => GetElement(this, index);
         set => this = WithElement(this, index, value);
@@ -69,49 +72,55 @@ public record struct AutoDiffTensor2<TN, U, TI>
 
     // Get
 
-    internal static Variable<TN, U> GetElement(AutoDiffTensor2<TN, U, TI> tensor, int index)
+    internal static TDN GetElement(AutoDiffTensor3<TDN, TN, U, TI> tensor, int index)
     {
-        if ((uint)index >= 2)
+        if ((uint)index >= 3)
             throw new IndexOutOfRangeException();
         return GetElementUnsafe(ref tensor, index);
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static Variable<TN, U> GetElementUnsafe(ref AutoDiffTensor2<TN, U, TI> tensor, int index)
+    private static TDN GetElementUnsafe(ref AutoDiffTensor3<TDN, TN, U, TI> tensor, int index)
     {
-        Debug.Assert(index is >= 0 and < 2);
-        return Unsafe.Add(ref Unsafe.As<AutoDiffTensor2<TN, U, TI>, Variable<TN, U>>(ref tensor), index);
+        Debug.Assert(index is >= 0 and < 3);
+        return Unsafe.Add(ref Unsafe.As<AutoDiffTensor3<TDN, TN, U, TI>, TDN>(ref tensor), index);
     }
 
     // Set
 
-    internal static AutoDiffTensor2<TN, U, TI> WithElement(AutoDiffTensor2<TN, U, TI> tensor, int index, Variable<TN, U> value)
+    internal static AutoDiffTensor3<TDN, TN, U, TI> WithElement(AutoDiffTensor3<TDN, TN, U, TI> tensor, int index, TDN value)
     {
-        if ((uint)index >= 2)
+        if ((uint)index >= 3)
             throw new IndexOutOfRangeException();
-        AutoDiffTensor2<TN, U, TI> result = tensor;
+        AutoDiffTensor3<TDN, TN, U, TI> result = tensor;
         SetElementUnsafe(ref result, index, value);
         return result;
     }
 
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private static void SetElementUnsafe(ref AutoDiffTensor2<TN, U, TI> tensor, int index, Variable<TN, U> value)
+    private static void SetElementUnsafe(ref AutoDiffTensor3<TDN, TN, U, TI> tensor, int index, TDN value)
     {
-        Debug.Assert(index is >= 0 and < 2);
-        Unsafe.Add(ref Unsafe.As<AutoDiffTensor2<TN, U, TI>, Variable<TN, U>>(ref tensor), index) = value;
+        Debug.Assert(index is >= 0 and < 3);
+        Unsafe.Add(ref Unsafe.As<AutoDiffTensor3<TDN, TN, U, TI>, TDN>(ref tensor), index) = value;
     }
-
-    //
-    // Formatting
-    //
-
-    public override readonly string ToString() => $"({X0}, {X1})";
 
     //
     // Methods
     //
 
-    public static AutoDiffTensor2<TN, U, TNI> Create<TNI>(Variable<TN, U> x0, Variable<TN, U> x1)
+    public static AutoDiffTensor3<TDN, TN, U, TI> Create(TN x0, TN x1, TN x3) => new(x0, x1, x3);
+
+    //
+    // Formatting
+    //
+
+    public override readonly string ToString() => $"({X0}, {X1}, {X2})";
+
+    //
+    // Methods
+    //
+
+    public static AutoDiffTensor3<TDN, TN, U, TNI> Create<TNI>(TDN x0, TDN x1, TDN x2)
         where TNI : IIndex
-        => new(x0, x1);
+        => new(x0, x1, x2);
 }
