+++ src/Mathematics.NET/DifferentialGeometry/Tensor`4.cs
@@ -35,136 +35,110 @@ using Mathematics.NET.Operations;
 
 namespace Mathematics.NET.DifferentialGeometry;
 
-/// <summary>Represents a rank-two tensor or a similar mathematical object.</summary>
-/// <typeparam name="TSM">A backing type that implements <see cref="ISquareMatrix{T, U, V, W}"/>.</typeparam>
-/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/>.</typeparam>
+/// <summary>Represents a rank-one tensor or a similar mathematical object.</summary>
+/// <typeparam name="TV">A backing type that implements <see cref="IVector{T, U, V, W}"/>.</typeparam>
+/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
-/// <typeparam name="TI1">The first index.</typeparam>
-/// <typeparam name="TI2">The second index.</typeparam>
-/// <param name="matrix">A backing matrix.</param>
+/// <typeparam name="TI">An index.</typeparam>
+/// <param name="vector">A backing vector.</param>
 [StructLayout(LayoutKind.Sequential)]
-public struct Tensor<TSM, TN, U, TI1, TI2>(TSM matrix)
-    : IRankTwoTensor<Tensor<TSM, TN, U, TI1, TI2>, TSM, TN, U, U, TI1, TI2>,
-      IAdditionOperation<Tensor<TSM, TN, U, TI1, TI2>, Tensor<TSM, TN, U, TI1, TI2>>,
-      ISubtractionOperation<Tensor<TSM, TN, U, TI1, TI2>, Tensor<TSM, TN, U, TI1, TI2>>,
-      IMultiplicationOperation<Tensor<TSM, TN, U, TI1, TI2>, TN, Tensor<TSM, TN, U, TI1, TI2>>,
-      IUnaryMinusOperation<Tensor<TSM, TN, U, TI1, TI2>, Tensor<TSM, TN, U, TI1, TI2>>,
-      IUnaryPlusOperation<Tensor<TSM, TN, U, TI1, TI2>, Tensor<TSM, TN, U, TI1, TI2>>
-    where TSM : ISquareMatrix<TSM, TN, U, U>
+public struct Tensor<TV, TN, U, TI>(TV vector)
+    : IRankOneTensor<Tensor<TV, TN, U, TI>, TV, TN, U, U, TI>,
+      IMultiplicationOperation<Tensor<TV, TN, U, TI>, TN, Tensor<TV, TN, U, TI>>,
+      IUnaryMinusOperation<Tensor<TV, TN, U, TI>, Tensor<TV, TN, U, TI>>,
+      IUnaryPlusOperation<Tensor<TV, TN, U, TI>, Tensor<TV, TN, U, TI>>
+    where TV : IVector<TV, TN, U, U>
     where TN : IComplex<TN, U, U>, IDifferentiableFunctions<TN>
     where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
-    where TI1 : IIndex
-    where TI2 : IIndex
+    where TI : IIndex
 {
-    private TSM _matrix = matrix;
-
-    //
-    // IRankTwoTensor Interface
-    //
-
-    public readonly IIndex I1 => TI1.Instance;
-
-    public readonly IIndex I2 => TI2.Instance;
+    private TV _vector = vector;
 
     //
     // IArrayRepresentable & Relevant Interfaces
     //
 
-    public static int Components => TSM.Components;
+    public static int Components => TV.Components;
 
-    public static int E1Components => TSM.E1Components;
+    public static int E1Components => TV.E1Components;
 
-    public static int E2Components => TSM.E2Components;
+    //
+    // IRankOneTensor Interface
+    //
+
+    public readonly IIndex I1 => TI.Instance;
 
     //
     // Indexer
     //
 
-    public TN this[int i, int j]
+    public TN this[int i]
     {
-        get => _matrix[i, j];
-        set => _matrix[i, j] = value;
+        get => _vector[i];
+        set => _vector[i] = value;
     }
 
     //
     // Operators
     //
 
-    public static Tensor<TSM, TN, U, TI1, TI2> operator -(Tensor<TSM, TN, U, TI1, TI2> tensor)
-        => new(-tensor._matrix);
+    public static Tensor<TV, TN, U, TI> operator -(Tensor<TV, TN, U, TI> tensor)
+        => new(-tensor._vector);
 
-    public static Tensor<TSM, TN, U, TI1, TI2> operator +(Tensor<TSM, TN, U, TI1, TI2> tensor)
+    public static Tensor<TV, TN, U, TI> operator +(Tensor<TV, TN, U, TI> tensor)
         => tensor;
 
-    public static Tensor<TSM, TN, U, TI1, TI2> operator +(Tensor<TSM, TN, U, TI1, TI2> left, Tensor<TSM, TN, U, TI1, TI2> right)
-        => left._matrix + right._matrix;
+    public static Tensor<TV, TN, U, TI> operator +(Tensor<TV, TN, U, TI> left, Tensor<TV, TN, U, TI> right)
+        => new(left._vector + right._vector);
 
-    public static Tensor<TSM, TN, U, TI1, TI2> operator -(Tensor<TSM, TN, U, TI1, TI2> left, Tensor<TSM, TN, U, TI1, TI2> right)
-        => left._matrix - right._matrix;
+    public static Tensor<TV, TN, U, TI> operator -(Tensor<TV, TN, U, TI> left, Tensor<TV, TN, U, TI> right)
+        => new(left._vector - right._vector);
 
-    public static Tensor<TSM, TN, U, TI1, TI2> operator *(TN c, Tensor<TSM, TN, U, TI1, TI2> tensor)
-        => new(c * tensor._matrix);
+    public static Tensor<TV, TN, U, TI> operator *(TN c, Tensor<TV, TN, U, TI> tensor)
+        => new(c * tensor._vector);
 
-    public static Tensor<TSM, TN, U, TI1, TI2> operator *(Tensor<TSM, TN, U, TI1, TI2> tensor, TN c)
-        => new(tensor._matrix * c);
+    public static Tensor<TV, TN, U, TI> operator *(Tensor<TV, TN, U, TI> tensor, TN c)
+        => new(tensor._vector * c);
 
     //
     // Equality
     //
 
-    public static bool operator ==(Tensor<TSM, TN, U, TI1, TI2> left, Tensor<TSM, TN, U, TI1, TI2> right)
-        => left._matrix == right._matrix;
+    public static bool operator ==(Tensor<TV, TN, U, TI> left, Tensor<TV, TN, U, TI> right)
+        => left._vector == right._vector;
 
-    public static bool operator !=(Tensor<TSM, TN, U, TI1, TI2> left, Tensor<TSM, TN, U, TI1, TI2> right)
-        => left._matrix != right._matrix;
+    public static bool operator !=(Tensor<TV, TN, U, TI> left, Tensor<TV, TN, U, TI> right)
+        => left._vector != right._vector;
 
-    public override readonly bool Equals([NotNullWhen(true)] object? obj) => obj is Tensor<TSM, TN, U, TI1, TI2> other && Equals(other);
+    public override readonly bool Equals([NotNullWhen(true)] object? obj) => obj is Tensor<TV, TN, U, TI> other && Equals(other);
 
-    public readonly bool Equals(Tensor<TSM, TN, U, TI1, TI2> value) => _matrix.Equals(value._matrix);
+    public readonly bool Equals(Tensor<TV, TN, U, TI> value) => _vector.Equals(value._vector);
 
-    public override readonly int GetHashCode() => HashCode.Combine(_matrix);
+    public override readonly int GetHashCode() => HashCode.Combine(_vector);
 
     //
     // Formatting
     //
 
-    public string ToString(string? format, IFormatProvider? provider) => _matrix.ToString(format, provider);
+    public string ToString(string? format, IFormatProvider? provider) => _vector.ToString(format, provider);
 
     //
     // Methods
     //
 
-    public TN[,] ToArray() => _matrix.ToArray();
+    public readonly TN[] ToArray() => _vector.ToArray();
 
-    /// <summary>Reinterpret this tensor as one with new indices.</summary>
-    /// <typeparam name="TNI1">A new first index.</typeparam>
-    /// <typeparam name="TNI2">A new second index.</typeparam>
-    /// <returns>A tensor with new indices.</returns>
+    /// <summary>Reinterpret this tensor as one with a new index.</summary>
+    /// <typeparam name="TNI">An index.</typeparam>
+    /// <returns>A tensor with a new index.</returns>
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public Tensor<TSM, TN, U, TNI1, TNI2> WithIndices<TNI1, TNI2>()
-        where TNI1 : IIndex
-        where TNI2 : IIndex
-        => Unsafe.As<Tensor<TSM, TN, U, TI1, TI2>, Tensor<TSM, TN, U, TNI1, TNI2>>(ref this);
-
-    /// <summary>Reinterpret this tensor as one with a new index in the first position.</summary>
-    /// <typeparam name="TNI">A new index.</typeparam>
-    /// <returns>A tensor with a new index in the first position.</returns>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public Tensor<TSM, TN, U, TNI, TI2> WithIndex1<TNI>()
+    public Tensor<TV, TN, U, TNI> WithIndex<TNI>()
         where TNI : IIndex
-        => Unsafe.As<Tensor<TSM, TN, U, TI1, TI2>, Tensor<TSM, TN, U, TNI, TI2>>(ref this);
-
-    /// <summary>Reinterpret this tensor as one with a new index in the second position.</summary>
-    /// <typeparam name="TNI">A new index.</typeparam>
-    /// <returns>A tensor with a new index in the second position.</returns>
-    [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    public Tensor<TSM, TN, U, TI1, TNI> WithIndex2<TNI>()
-        where TNI : IIndex
-        => Unsafe.As<Tensor<TSM, TN, U, TI1, TI2>, Tensor<TSM, TN, U, TI1, TNI>>(ref this);
+        => Unsafe.As<Tensor<TV, TN, U, TI>, Tensor<TV, TN, U, TNI>>(ref this);
 
     //
     // Implicit Operators
     //
 
-    public static implicit operator Tensor<TSM, TN, U, TI1, TI2>(TSM input) => new(input);
+    public static implicit operator Tensor<TV, TN, U, TI>(TV input) => new(input);
 }
