+++ src/Mathematics.NET/Solvers/RungeKutta4`5.cs
@@ -1,4 +1,4 @@
-// <copyright file="RungeKutta4`3.cs" company="Mathematics.NET">
+// <copyright file="RungeKutta4`4.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -28,36 +28,34 @@
 using System.Numerics;
 using System.Runtime.CompilerServices;
 using CommunityToolkit.HighPerformance.Helpers;
+using Mathematics.NET.DifferentialGeometry.Abstractions;
 using Mathematics.NET.LinearAlgebra.Abstractions;
 using Mathematics.NET.Operations;
 
 namespace Mathematics.NET.Solvers;
 
 /// <summary>Represents a fourth-order Runge-Kutta solver.</summary>
-/// <typeparam name="TSI">A type that implements <see cref="IStateItem{TSC, TA, TN, TB}"/>.</typeparam>
-/// <typeparam name="TA">An array-like object that supports addition and multiplication on its elements.</typeparam>
+/// <typeparam name="TR1T">A rank-one tensor.</typeparam>
+/// <typeparam name="TV">The backing type of the tensor.</typeparam>
 /// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
 /// <typeparam name="TB">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
+/// <typeparam name="TI">The index of the tensor.</typeparam>
 /// <param name="function">A function to use for the integration step.</param>
-public sealed class RungeKutta4<TSI, TA, TN, TB>(Func<TN, TSI, TSI> function)
-    where TSI : IStateItem<TSI, TA, TN, TB>
-    where TA
-    : I1DArrayRepresentable<TA, TN, TB, TB>,
-      IAdditionOperation<TA, TA>,
-      ISubtractionOperation<TA, TA>,
-      IMultiplicationOperation<TA, TN, TA>,
-      IUnaryMinusOperation<TA, TA>
+public sealed class RungeKutta4<TR1T, TV, TN, TB, TI>(Func<TN, TR1T, TR1T> function)
+    where TR1T : IRankOneTensor<TR1T, TV, TN, TB, TB, TI>, IMultiplicationOperation<TR1T, TN, TR1T>
+    where TV : IVector<TV, TN, TB, TB>
     where TN : IComplex<TN, TB, TB>, IDifferentiableFunctions<TN>
     where TB : IBinaryFloatingPointIeee754<TB>, IMinMaxValue<TB>
+    where TI : IIndex
 {
-    private readonly struct RK4IntegrateAction(Func<TN, TSI, TSI> function, TN time, TN dt) : IRefAction<TSI>
+    private readonly struct RK4IntegrateAction(Func<TN, TR1T, TR1T> function, TN time, TN dt) : IRefAction<TR1T>
     {
-        private readonly Func<TN, TSI, TSI> _function = function;
+        private readonly Func<TN, TR1T, TR1T> _function = function;
         private readonly TN _time = time;
         private readonly TN _dt = dt;
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public void Invoke(ref TSI value)
+        public void Invoke(ref TR1T value)
         {
             var k1 = _function(_time, value);
             var k2 = _function(_time + IBinaryFloatingPointIeee754<TB>.Half * _dt, value + IBinaryFloatingPointIeee754<TB>.Half * k1 * _dt);
@@ -67,10 +65,10 @@ public sealed class RungeKutta4<TSI, TA, TN, TB>(Func<TN, TSI, TSI> function)
         }
     }
 
-    private readonly Func<TN, TSI, TSI> _function = function;
+    private readonly Func<TN, TR1T, TR1T> _function = function;
 
     /// <inheritdoc cref="RungeKutta4{T, U}.Integrate(State{T, U}, T)"/>
-    public void Integrate(State<TSI, TA, TN, TB> state, TN dt)
+    public void Integrate(State<TR1T, TV, TN, TB, TI> state, TN dt)
     {
         var system = state.System.Span;
         var time = state.Time;
@@ -87,7 +85,7 @@ public sealed class RungeKutta4<TSI, TA, TN, TB>(Func<TN, TSI, TSI> function)
     }
 
     /// <inheritdoc cref="RungeKutta4{T, U}.IntegrateParallel(State{T, U}, T)"/>
-    public void IntegrateParallel(State<TSI, TA, TN, TB> state, TN dt)
+    public void IntegrateParallel(State<TR1T, TV, TN, TB, TI> state, TN dt)
     {
         ParallelHelper.ForEach(state.System, new RK4IntegrateAction(_function, state.Time, dt));
         state.Time += dt;
