+++ src/Mathematics.NET/Solvers/State`2.cs
@@ -1,4 +1,4 @@
-// <copyright file="State`2.cs" company="Mathematics.NET">
+// <copyright file="State`1.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -26,24 +26,21 @@
 // </copyright>
 
 using System.Numerics;
-using Mathematics.NET.LinearAlgebra.Abstractions;
 
 namespace Mathematics.NET.Solvers;
 
 /// <summary>Represents the state of a system.</summary>
-/// <typeparam name="TV">A type that implements <see cref="IVector{T, U, V, W}"/>.</typeparam>
-/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
-/// <typeparam name="TB">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
+/// <typeparam name="T">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
+/// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
 /// <param name="system">The system.</param>
 /// <param name="time">The time.</param>
-public sealed class State<TV, TN, TB>(Memory<TV> system, TN time)
-    where TV : IVector<TV, TN, TB, TB>
-    where TN : IComplex<TN, TB, TB>, IDifferentiableFunctions<TN>
-    where TB : IBinaryFloatingPointIeee754<TB>, IMinMaxValue<TB>
+public sealed class State<T, U>(Memory<T> system, T time)
+    where T : IComplex<T, U, U>, IDifferentiableFunctions<T>
+    where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
 {
-    /// <inheritdoc cref="State{T, U}.System"/>
-    public Memory<TV> System = system;
+    /// <summary>The system.</summary>
+    public Memory<T> System = system;
 
-    /// <inheritdoc cref="State{T, U}.Time"/>
-    public TN Time = time;
+    /// <summary>The time.</summary>
+    public T Time = time;
 }
