+++ src/Mathematics.NET/Solvers/State`5.cs
@@ -1,4 +1,4 @@
-// <copyright file="State`1.cs" company="Mathematics.NET">
+// <copyright file="State`4.cs" company="Mathematics.NET">
 // Mathematics.NET
 // https://github.com/HamletTanyavong/Mathematics.NET
 //
@@ -26,21 +26,29 @@
 // </copyright>
 
 using System.Numerics;
+using Mathematics.NET.DifferentialGeometry.Abstractions;
+using Mathematics.NET.LinearAlgebra.Abstractions;
 
 namespace Mathematics.NET.Solvers;
 
 /// <summary>Represents the state of a system.</summary>
-/// <typeparam name="T">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
-/// <typeparam name="U">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
+/// <typeparam name="TR1T">A rank-one tensor.</typeparam>
+/// <typeparam name="TV">The backing type of the tensor.</typeparam>
+/// <typeparam name="TN">A type that implements <see cref="IComplex{T, U, V}"/> and <see cref="IDifferentiableFunctions{T}"/>.</typeparam>
+/// <typeparam name="TB">A type that implements <see cref="IBinaryFloatingPointIeee754{TSelf}"/> and <see cref="IMinMaxValue{TSelf}"/>.</typeparam>
+/// <typeparam name="TI">The index of the tensor.</typeparam>
 /// <param name="system">The system.</param>
 /// <param name="time">The time.</param>
-public sealed class State<T, U>(Memory<T> system, T time)
-    where T : IComplex<T, U, U>, IDifferentiableFunctions<T>
-    where U : IBinaryFloatingPointIeee754<U>, IMinMaxValue<U>
+public sealed class State<TR1T, TV, TN, TB, TI>(Memory<TR1T> system, TN time)
+    where TR1T : IRankOneTensor<TR1T, TV, TN, TB, TB, TI>
+    where TV : IVector<TV, TN, TB, TB>
+    where TN : IComplex<TN, TB, TB>, IDifferentiableFunctions<TN>
+    where TB : IBinaryFloatingPointIeee754<TB>, IMinMaxValue<TB>
+    where TI : IIndex
 {
-    /// <summary>The system.</summary>
-    public Memory<T> System = system;
+    /// <inheritdoc cref="State{T, U}.System"/>
+    public Memory<TR1T> System = system;
 
-    /// <summary>The time.</summary>
-    public T Time = time;
+    /// <inheritdoc cref="State{T, U}.Time"/>
+    public TN Time = time;
 }
